api/stream: split types.js into proxy.js and ffmpeg.js
